product create split to file
